feature(async): added 'next-router-mock/async' entrypoint, allowing for sync and async versions of the router
